SelectFromTable изменён интерфейс

+++ Kvit_main/table_tariff.h
@@ -12,7 +12,7 @@ public:
     table_tariff();
     ~table_tariff();
 
-    float is_Tariff(int usluga, QDate date, int tarif = 1);           //возвращает размер тарифа
+    double is_Tariff(int usluga, QDate date, int tarif = 1);           //возвращает размер тарифа
     QSqlQueryModel* ModelTarifTabl(int year,int month);                         //тарифы на месяц
     void UpdateTarif(QString tarif,QString tarif2, QString norma, int idusl);
     int FillTarif(int month, int year);
